more refactoring, trying to rename project to more accurate title

+++ src/main.cpp
@@ -1,18 +1,18 @@
-#include "Classes/Singleton/Game/Game.h"
+#include "Singleton/Game/Game.h"
 
 int main() {
     SDL_SetMainReady();
 
     while (Game::getInstance().isRunning()) {
-        const Uint32 frameStart = SDL_GetTicks64();
+        const Uint64 frameStart = SDL_GetTicks64();
 
         Game::getInstance().handleEvents();
         Game::getInstance().update();
         Game::getInstance().render();
 
-        const Uint32 frameTime = SDL_GetTicks64() - frameStart;
+        const Uint64 frameTime = SDL_GetTicks64() - frameStart;
 
-        if (const Uint32 frameDelay = 1000 / Game::getInstance().getFps(); frameDelay > frameTime) {
+        if (const Uint64 frameDelay = 1000 / Game::getInstance().getFps(); frameDelay > frameTime) {
             SDL_Delay(frameDelay - frameTime);
         }
     }
